backend/requirements.txt
click>=8.1.0
pypdf>=4.0.0
pdfplumber>=0.10.0
requests>=2.31.0
beautifulsoup4>=4.12.0
sqlalchemy>=2.0.0
pandas>=2.0.0
rich>=13.7.0
yfinance>=0.2.36
lxml>=4.9.0
python-dotenv>=1.0.0
tabulate>=0.9.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
